Upgrade 0.1.1 (#20)
* Fix warn to double to long double for Clang

* Update xmaho version to 0.1.1

* Fix documentation

* Add example codes to point

* Add documentation for point::norm

* Generate docs

* Fix URL for API

* Add example for std_ext/valarray.hpp

* Regenerate docs

+++ test/std_ext/valarray.cpp
@@ -43,7 +43,7 @@ template<typename T>
 auto get_uniform_distribution(std::size_t data_count)
 {
   constexpr auto reciprocal {1. / (testing_max_norm + 1)};
-  const auto max_value {std::pow(std::numeric_limits<T>::max() / static_cast<long double>(data_count), reciprocal)};
+  const auto max_value {std::pow(static_cast<long double>(std::numeric_limits<T>::max()) / static_cast<long double>(data_count), reciprocal)};
   EXPECT_GT(std::pow(std::numeric_limits<T>::max(), reciprocal), 2);
   EXPECT_GT(max_value, 1);
